backend/requirements.txt
django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels[daphne]
channels-redis
celery[redis]
redis
psycopg2-binary
scapy
geoip2
scikit-learn
joblib
numpy
pandas
weasyprint
django-filter
Pillow
requests
python-decouple
psutil
stix2
